Storage/requirements.txt
connexion==2.14.2
swagger-ui-bundle==0.0.9
SQLAlchemy>=1.4.0,<3.0.0
mysql-connector-python>=8.0.0,<9.0.0
pyMySQL>=1.0.0,<2.0.0
pykafka==2.8.0